move to_integer_class to upolybase.h
